Add automatic model conversion to onnx, ncnn and ave-tracker bundle (#43)
This PR brings the functionality for model conversion inspired by the avelab model conversion pipeline used in face tracking.

The converter takes as input the model folder with pb files and variable folder from a training and converts it into several onnx files. If ave-tracker path is given then the models are converted into ncnn files (optimized and binarized), a bundle file to be used in ave-tracker containing a settings.json and the ncnn file is also generated. The settings.json file is automatically created. Furthermore a info.txt file is created with md5 checksum of the relevant files used as inputs.